Added permissions to different requests in Events

+++ src/main/java/com/az/mykowel/controller/EventsController.java
@@ -9,7 +9,9 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import com.az.mykowel.model.entity.Events;
+import com.az.mykowel.model.entity.Users;
 import com.az.mykowel.model.services.EventsService;
+import com.az.mykowel.model.services.UserService;
 
 @RestController
 @RequestMapping("events")
@@ -17,6 +19,9 @@ public class EventsController {
     
     @Autowired
     private EventsService eventsService;
+
+    @Autowired
+    private UserService userService;
 
 
     @GetMapping("/get")
@@ -29,20 +34,32 @@ public class EventsController {
     }
 
     @PostMapping(value = "/add", consumes = {"*/*"})
-    public ResponseEntity<?> add(@ModelAttribute Events event){
+    public ResponseEntity<?> add(@ModelAttribute Events event, @RequestHeader String token){
         try{
-            eventsService.save(event);
-            return new ResponseEntity<>("Event added", HttpStatus.OK);
+            Users user = userService.getUserByToken(token);
+            if(user.checkPerms(user.getIs_admin())) {
+                eventsService.save(event);
+                return new ResponseEntity<>("Event added", HttpStatus.OK);
+            }
+            else{
+                return new ResponseEntity<>("Permissions denied", HttpStatus.FORBIDDEN);
+            }
         }catch(NoSuchElementException e){
             return new ResponseEntity<>("Not Found", HttpStatus.NOT_FOUND); 
         }
     }
 
     @DeleteMapping(value = "/delete/{Id}")
-    public ResponseEntity<?> delete(@PathVariable Long Id){
+    public ResponseEntity<?> delete(@PathVariable Long Id, @RequestHeader String token){
         try{
-            eventsService.delete(Id);
-            return new ResponseEntity<>("Event deleted", HttpStatus.OK);
+            Users user = userService.getUserByToken(token);
+            if(user.checkPerms(user.getIs_admin())) {
+                eventsService.delete(Id);
+                return new ResponseEntity<>("Event deleted", HttpStatus.OK);
+            }
+            else{
+                return new ResponseEntity<>("Permissions denied", HttpStatus.FORBIDDEN);
+            }
         }catch(NoSuchElementException e){
             return new ResponseEntity<>("Not Found", HttpStatus.NOT_FOUND); 
         }
